Clean up bullet point explaining using keys instead of useEffect for resetting a component
## Because
This bullet point is confusingly worded. 


## This PR
* Reworded this sentence into two sentences for clarity:
"Just like using a key on a list’s item, adding one to a component, based on the state on which it should be reset creates a unique version of that component for each change in the value of the state."


## Issue
No pull request for this.

## Additional Information
No additional information.